pyproject.toml
[tool.poetry]
name = "django-ratelimiter"
version = "0.1.1"
description = "Rate-limiting for django"
authors = ["Andrii Kohut <kogut.andriy@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
django = "*"
limits = ">=3.10"
typing-extensions = {version = "^4.11.0", python = "3.9"}


[tool.poetry.group.dev.dependencies]
django-stubs = "^4.2.7"
mypy = "^1.9.0"
black = "^24.4.0"
ruff = "^0.3.5"
pytest = "^8.1.1"
freezegun = "^1.4.0"
pymemcache = "^4.0.0"
redis = "^5.0.3"
pytest-cov = "^5.0.0"
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.18"
djangorestframework = "^3.15.1"
django-ninja = "^1.1.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test_app.settings"
